Extract to src/components/managerlib/bibwatcher.ts
